requirements.txt
Pillow
fpdf
pytesseract
opencv-python